test: move mime parsing unit tests to dc_mimeparser.rs
